+++ main.py
@@ -299,7 +299,7 @@ if (mode_selection == "0" and config["general"]["modes"]["enabled"]["management"
                 pass # Do nothing, and just finish this loop.
             elif (selection == "1"): # The user has selected the "view files" option.
                 os.system("ls -1 " + config["general"]["working_directory"]) # Run the 'ls' command in the working directory.
-                prompt("Press enter to continue...", optional=True, input_type=str, default="") # Wait for the user to press enter before continuing
+                prompt(style.faint + "\nPress enter to continue..." + style.end, optional=True, input_type=str, default="") # Wait for the user to press enter before repeating the menu loop.
             elif (selection == "2"): # The user has selected the "copy files" option.
 
                 # Reset all of the file selections to un-selected.
@@ -577,6 +577,8 @@ if (mode_selection == "0" and config["general"]["modes"]["enabled"]["management"
                     display_message("The disk usage could not be displayed because the 'disk_statistics' configuration option is disabled.", 2)
             else: # The user has selected an invalid option in the information menu.
                 display_message("Invalid selection.", 2)
+
+            prompt(style.faint + "\nPress enter to continue..." + style.end, optional=True, input_type=str, default="") # Wait for the user to press enter before repeating the menu loop.
             
 
 
@@ -980,7 +982,7 @@ elif (mode_selection == "1" and config["general"]["modes"]["enabled"]["prerecord
 
 
     debug_message("Waiting for user input")
-    prompt("Press enter to continue...", optional=True, input_type=str, default="") # Wait for the user to press enter before continuing
+    prompt(style.faint + "\nPress enter to continue..." + style.end, optional=True, input_type=str, default="") # Wait for the user to press enter before repeating the menu loop.
 
     debug_message("Starting menu loop")
     while True: # Run the pre-recorded mode menu in a loop forever until the user exits.
@@ -1103,7 +1105,7 @@ elif (mode_selection == "1" and config["general"]["modes"]["enabled"]["prerecord
                 else:
                     display_message("Invalid selection.", 2)
 
-            prompt("\nPress enter to continue...", optional=True, input_type=str, default="") # Wait for the user to press enter before repeating the menu loop.
+            prompt(style.faint + "\nPress enter to continue..." + style.end, optional=True, input_type=str, default="") # Wait for the user to press enter before repeating the menu loop.
 
 
         elif (selection == "2"): # The user has selected to manage object recognition data.
@@ -1152,7 +1154,7 @@ elif (mode_selection == "1" and config["general"]["modes"]["enabled"]["prerecord
             else: # The user has selected the object recognition data management menu, but object recognition has been disabled.
                 display_message("Object recognition has been disabled. There is no object recogntion data to manage.", 2)
 
-            prompt("\nPress enter to continue...", optional=True, input_type=str, default="") # Wait for the user to press enter before repeating the menu loop.
+            prompt(style.faint + "\nPress enter to continue..." + style.end, optional=True, input_type=str, default="") # Wait for the user to press enter before repeating the menu loop.
 
 
         elif (selection == "3"): # The user has selected to manage GPX location information.
@@ -1203,7 +1205,7 @@ elif (mode_selection == "1" and config["general"]["modes"]["enabled"]["prerecord
             else:
                 display_message("GPX processing has been disabled since a GPX file wasn't provided. There is no GPX location data to manage.", 2)
 
-            prompt("\nPress enter to continue...", optional=True, input_type=str, default="") # Wait for the user to press enter before repeating the menu loop.
+            prompt(style.faint + "\nPress enter to continue..." + style.end, optional=True, input_type=str, default="") # Wait for the user to press enter before repeating the menu loop.
 
 
         elif (selection == "4"): # If the user selects option 4 on the main menu, then show the statstics for this session.
@@ -1211,12 +1213,12 @@ elif (mode_selection == "1" and config["general"]["modes"]["enabled"]["prerecord
             print("    Plates found: " + str(len(plates_detected))) # Show how many unique plates were detected.
             print("    Videos analyzed: " + str(len(videos))) # Show how many videos were analyzed.
             print("    Alerts detected: " + str(len(active_alerts))) # Show how many videos were analyzed.
-            prompt("\nPress enter to continue...", optional=True, input_type=str, default="") # Wait for the user to press enter before repeating the menu loop.
+            prompt(style.faint + "\nPress enter to continue..." + style.end, optional=True, input_type=str, default="") # Wait for the user to press enter before repeating the menu loop.
 
 
         else: # If the user selects an unrecognized option on the main menu for pre-recorded mode, then show a warning.
             display_message("Invalid selection.", 2)
-            prompt("\nPress enter to continue...", optional=True, input_type=str, default="") # Wait for the user to press enter before repeating the menu loop.
+            prompt(style.faint + "\nPress enter to continue..." + style.end, optional=True, input_type=str, default="") # Wait for the user to press enter before repeating the menu loop.